# Vectorize BLAKE2b