--- a/src/util/mod.rs
+++ b/src/util/mod.rs
@@ -25,6 +25,7 @@ use subtle::ConstantTimeEq;
 
 pub(crate) mod endianness;
 pub(crate) mod u32x4;
+pub(crate) mod u64x4;
 
 #[must_use = "SECURITY WARNING: Ignoring a Result can have real security implications."]
 #[cfg(feature = "safe_api")]