This changes the internal BLAKE2b permutation to resemble SIMD vectorization (similar to how ChaCha20 [was vectorized](https://github.com/brycx/orion/pull/84)).

The approach to this vectorization has been taken from the [`blake2-rfc`](https://github.com/cesarb/blake2-rfc) crate.

This has shown the following increases in performance (measured on a Intel Core i7-4790K 4.0 GHz):

```
BLAKE2b-512/compute hash/512                                                                             
                        time:   [609.79 ns 610.74 ns 611.70 ns]
                        thrpt:  [798.24 MiB/s 799.50 MiB/s 800.73 MiB/s]
                 change:
                        time:   [-5.4893% -4.5062% -3.1013%] (p = 0.00 < 0.05)
                        thrpt:  [+3.2005% +4.7188% +5.8082%]
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  3 (3.00%) low mild
  2 (2.00%) high mild
  6 (6.00%) high severe
BLAKE2b-512/compute hash/1024                                                                             
                        time:   [1.1434 us 1.1449 us 1.1465 us]
                        thrpt:  [851.77 MiB/s 852.97 MiB/s 854.12 MiB/s]
                 change:
                        time:   [-4.1132% -3.4438% -2.8255%] (p = 0.00 < 0.05)
                        thrpt:  [+2.9077% +3.5666% +4.2896%]
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) low mild
  1 (1.00%) high mild
  3 (3.00%) high severe
BLAKE2b-512/compute hash/2048                                                                             
                        time:   [2.2060 us 2.2097 us 2.2136 us]
                        thrpt:  [882.32 MiB/s 883.87 MiB/s 885.36 MiB/s]
                 change:
                        time:   [-3.8288% -3.1037% -2.4512%] (p = 0.00 < 0.05)
                        thrpt:  [+2.5128% +3.2031% +3.9812%]
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) high mild
  3 (3.00%) high severe
BLAKE2b-512/compute hash/4098                                                                             
                        time:   [4.5469 us 4.5549 us 4.5633 us]
                        thrpt:  [856.44 MiB/s 858.02 MiB/s 859.51 MiB/s]
                 change:
                        time:   [-3.6534% -3.0408% -2.4649%] (p = 0.00 < 0.05)
                        thrpt:  [+2.5272% +3.1362% +3.7919%]
```